Bruk insert og sjekk i database om sykmelding finnes fra før
