+++ src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/SykmeldingRepositoryTest.kt
@@ -9,7 +9,6 @@ import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.time.LocalDate
 import java.util.UUID
 
 class SykmeldingRepositoryTest {
@@ -33,28 +32,6 @@ class SykmeldingRepositoryTest {
     }
 
     @Test
-    fun `lagreSykmelding skal ignorere sykmeldinger med sykmelding-IDer som allerede er lagret`() {
-        val nowDate = LocalDate.now()
-        val sykmeldinger =
-            List(2) { nowDate }.mapIndexed { index, now ->
-                sykmeldingMock().copyWithSyketilfelleStartDato(
-                    now.minusDays(
-                        index.toLong(),
-                    ),
-                )
-            }
-
-        sykmeldinger[0].lagreSykmelding(sykmeldingRepository)
-        sykmeldinger[1].lagreSykmelding(sykmeldingRepository)
-
-        val lagredeSykmeldinger =
-            transaction(db) { SykmeldingEntitet.selectAll().mapNotNull { it.getOrNull(arbeidsgiverSykmelding) } }
-
-        lagredeSykmeldinger.size shouldBe 1
-        lagredeSykmeldinger[0].syketilfelleStartDato shouldBe sykmeldinger[0].sykmelding.syketilfelleStartDato
-    }
-
-    @Test
     fun `hentSykmelding skal hente sykmelding med id`() {
         val sykmeldinger = List(10) { UUID.randomUUID() }.map { id -> sykmeldingMock().copyId(id.toString()) }
 
@@ -67,9 +44,6 @@ class SykmeldingRepositoryTest {
 }
 
 private fun SendSykmeldingAivenKafkaMessage.copyId(id: String) = copy(sykmelding = sykmelding.copy(id = id))
-
-private fun SendSykmeldingAivenKafkaMessage.copyWithSyketilfelleStartDato(syketilfelleStartDato: LocalDate) =
-    copy(sykmelding = sykmelding.copy(syketilfelleStartDato = syketilfelleStartDato))
 
 private fun SendSykmeldingAivenKafkaMessage.lagreSykmelding(sykmeldingRepository: SykmeldingRepository) {
     sykmeldingRepository.lagreSykmelding(
